WAITM-165 Move api fetch outside of imaging request printout
